Server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Rahul Tarak <cryogenicplanet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
opencv-python-headless = "4.9.0.80"
uvicorn = "0.15.0"
Pillow = "^8.3.1"
requests = "2.26.0"
torch = "^2.2.1"
transformers = "4.39.1"
gradio = "^4.22.0"
ultralytics = "^8.1.33"
einops = "^0.7.0"
timm = "^0.9.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
